requirements.txt
# --- Web framework ---
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.36
werkzeug==3.0.4
gunicorn==23.0.0
python-dotenv==1.0.1

# --- RAG stack ---
langchain==0.3.7
langchain-core==0.3.21
langchain-community==0.3.7
langchain-ollama==0.2.0
langchain-chroma==0.1.4
chromadb==0.5.20

# --- NLP tiếng Việt ---
underthesea==6.8.4
unidecode==1.3.8

# --- Misc ---
markdown==3.7
requests==2.32.3
